+++ src/test/java/de/adorsys/keycloak/config/service/ImportRolesIT.java
@@ -28,6 +28,7 @@ import de.adorsys.keycloak.config.util.VersionUtil;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
+import org.keycloak.representations.idm.ClientRepresentation;
 import org.keycloak.representations.idm.RealmRepresentation;
 import org.keycloak.representations.idm.RoleRepresentation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -861,10 +862,10 @@ class ImportRolesIT extends AbstractImportTest {
     }
 
     @Test
-    @Order(30)
+    @Order(70)
     @SuppressWarnings("deprecation")
     void shouldNotThrowImportingClientRoleThatAlreadyExists() throws IOException {
-        RealmImport foundImport = getFirstImport("30_import_realm_with_duplicated_client_role.json");
+        RealmImport foundImport = getFirstImport("70_import_realm_with_duplicated_client_role.json");
 
         assertThat(
                 foundImport
@@ -888,8 +889,172 @@ class ImportRolesIT extends AbstractImportTest {
         assertDoesNotThrow(() -> realmImportService.doImport(foundImport));
     }
 
+    @Test
+    @Order(71)
+    void shouldImportRealmWithNestedComposites() throws IOException {
+        doImport("71.1_import_realm_with_nested_composites.json");
+
+        String REALM_NAME = "realmWithRoles71";
+
+        RealmRepresentation realm = keycloakProvider.getInstance().realm(REALM_NAME).partialExport(true, true);
+
+        assertThat(realm.getRealm(), is(REALM_NAME));
+
+        RoleRepresentation realmRole;
+        RoleRepresentation.Composites composites;
+
+        realmRole = keycloakRepository.getRealmRole(
+                realm, "subscription_user"
+        );
+
+        assertThat(realmRole.getName(), is("subscription_user"));
+        assertThat(realmRole.getDescription(), is("subscription"));
+        assertThat(realmRole.isComposite(), is(true));
+        assertThat(realmRole.getClientRole(), is(false));
+
+        composites = realmRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), hasEntry(is("fe"), containsInAnyOrder("subscription_user")));
+        assertThat(composites.getClient(), hasEntry(is("be"), containsInAnyOrder("subscription_user")));
+        assertThat(composites.getClient(), aMapWithSize(2));
+        assertThat(composites.getRealm(), is(nullValue()));
+
+
+        realmRole = keycloakRepository.getRealmRole(
+                realm, "procurement_user"
+        );
+
+        assertThat(realmRole.getName(), is("procurement_user"));
+        assertThat(realmRole.getDescription(), is("procurement"));
+        assertThat(realmRole.isComposite(), is(true));
+        assertThat(realmRole.getClientRole(), is(false));
+
+        composites = realmRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), hasEntry(is("fe"), containsInAnyOrder("procurement_user")));
+        assertThat(composites.getClient(), hasEntry(is("be"), containsInAnyOrder("procurement_user")));
+        assertThat(composites.getClient(), aMapWithSize(2));
+        assertThat(composites.getRealm(), is(nullValue()));
+
+
+        realmRole = keycloakRepository.getRealmRole(
+                realm, "vendor_user"
+        );
+
+        assertThat(realmRole.getName(), is("vendor_user"));
+        assertThat(realmRole.getDescription(), is("vendor"));
+        assertThat(realmRole.isComposite(), is(true));
+        assertThat(realmRole.getClientRole(), is(false));
+
+        composites = realmRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), hasEntry(is("fe"), containsInAnyOrder("vendor_user")));
+        assertThat(composites.getClient(), hasEntry(is("be"), containsInAnyOrder("vendor_user")));
+        assertThat(composites.getClient(), aMapWithSize(2));
+        assertThat(composites.getRealm(), is(nullValue()));
+
+
+        realmRole = keycloakRepository.getRealmRole(
+                realm, "default-roles-realmWithRoles71"
+        );
+
+        assertThat(realmRole.getName(), is("default-roles-realmWithRoles71"));
+        assertThat(realmRole.getDescription(), is("${role_default-roles}"));
+        assertThat(realmRole.isComposite(), is(true));
+        assertThat(realmRole.getClientRole(), is(false));
+
+        composites = realmRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), hasEntry(is("account"), containsInAnyOrder("view-profile", "manage-account")));
+        assertThat(composites.getClient(), aMapWithSize(1));
+        assertThat(composites.getRealm(), containsInAnyOrder("offline_access", "subscription_user", "uma_authorization"));
+
+
+        List<String> userRealmLevelRoles = keycloakRepository.getUserRealmLevelRoles(
+                REALM_NAME, "user@test.com"
+        );
+
+        assertThat(userRealmLevelRoles, hasItem("subscription_user"));
+
+        doImport("71.2_import_realm_with_nested_composites.json");
+
+        realm = keycloakProvider.getInstance().realm(REALM_NAME).partialExport(true, true);
+
+        realmRole = keycloakRepository.getRealmRole(
+                realm, "subscription_user"
+        );
+
+        assertThat(realmRole.getName(), is("subscription_user"));
+        assertThat(realmRole.getDescription(), is("subscription"));
+        assertThat(realmRole.isComposite(), is(true));
+        assertThat(realmRole.getClientRole(), is(false));
+
+        composites = realmRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), hasEntry(is("fe"), containsInAnyOrder("procurement_user")));
+        assertThat(composites.getClient(), hasEntry(is("fe2"), containsInAnyOrder("subscription_user")));
+        assertThat(composites.getClient(), aMapWithSize(2));
+        assertThat(composites.getRealm(), is(nullValue()));
+
+
+        realmRole = keycloakRepository.getRealmRole(
+                realm, "procurement_user"
+        );
+
+        assertThat(realmRole.getName(), is("procurement_user"));
+        assertThat(realmRole.getDescription(), is("procurement"));
+        assertThat(realmRole.isComposite(), is(true));
+        assertThat(realmRole.getClientRole(), is(false));
+
+        composites = realmRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), hasEntry(is("be2"), containsInAnyOrder("subscription_user")));
+        assertThat(composites.getClient(), hasEntry(is("be3"), containsInAnyOrder("procurement_user")));
+        assertThat(composites.getClient(), aMapWithSize(2));
+        assertThat(composites.getRealm(), is(nullValue()));
+
+
+        realmRole = keycloakRepository.getRealmRole(
+                realm, "vendor_user"
+        );
+
+        assertThat(realmRole.getName(), is("vendor_user"));
+        assertThat(realmRole.getDescription(), is("vendor"));
+        assertThat(realmRole.isComposite(), is(true));
+        assertThat(realmRole.getClientRole(), is(false));
+
+        composites = realmRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), hasEntry(is("fe"), containsInAnyOrder("vendor_user")));
+        assertThat(composites.getClient(), hasEntry(is("be2"), containsInAnyOrder("vendor_user")));
+        assertThat(composites.getClient(), aMapWithSize(2));
+        assertThat(composites.getRealm(), is(nullValue()));
+
+
+        doImport("71.3_import_realm_with_nested_composites.json");
+
+        realm = keycloakProvider.getInstance().realm(REALM_NAME).partialExport(true, true);
+
+        RoleRepresentation ClientRole = keycloakRepository.getClientRole(
+                realm, "fe", "composite_role_user"
+        );
+
+        assertThat(ClientRole.getName(), is("composite_role_user"));
+        assertThat(ClientRole.getDescription(), is("composite role created BEFORE dependent roles have been created."));
+        assertThat(ClientRole.isComposite(), is(true));
+        assertThat(ClientRole.getClientRole(), is(true));
+
+        composites = ClientRole.getComposites();
+        assertThat(composites, notNullValue());
+        assertThat(composites.getClient(), aMapWithSize(2));
+        assertThat(composites.getClient(), hasEntry(is("be"), containsInAnyOrder("subscription_user", "procurement_user", "vendor_user")));
+        assertThat(composites.getClient(), hasEntry(is("fe"), containsInAnyOrder("subscription_user", "procurement_user", "vendor_user")));
+        assertThat(composites.getRealm(), is(nullValue()));
+    }
+
     @SuppressWarnings("SpringJavaAutowiredMembersInspection")
     @Nested
+    @Order(60)
     @TestPropertySource(properties = {
             "import.remove-default-role-from-user=true"
     })
@@ -898,9 +1063,9 @@ class ImportRolesIT extends AbstractImportTest {
         public RealmImportService realmImportService;
 
         @Test
-        @Order(82)
+        @Order(0)
         void shouldCreateUserAndRemoveDefaultRole() throws IOException {
-            doImport("80_update_realm__add_user_with_realm_role.json", realmImportService);
+            doImport("60_update_realm__add_user_with_realm_role.json", realmImportService);
 
             RealmRepresentation realm = keycloakProvider.getInstance().realm(REALM_NAME).partialExport(true, true);